Implement "top-down" triplet searching.
Previously, the algorithm used "bottom-up" triplet searching, in which triplets were formed by finding the best corresponding mixture eigenvector for each possible pair of independent eigenvectors. Now, to ensure that the set of triplets covers every eigenvector exactly once as a mixture eigenvector, we search for the best possible pair of independent eigenvectors that form each eigenvector as a mixture.